TST: tseries/tests/test_timeseries.py tests moved to appropriate places.
